Rename lr_r2 to LR2
