+++ py-utils/test/discovery/mocked_health_gen/lr2/mocked_node_health.json
